Add Ford API support
