+++ vehicle/ford.go
@@ -1,11 +1,9 @@
 package vehicle
 
 import (
-	"errors"
 	"fmt"
 	"net/http"
 	"net/url"
-	"strconv"
 	"strings"
 	"time"
 
@@ -13,6 +11,7 @@ import (
 	"github.com/andig/evcc/provider"
 	"github.com/andig/evcc/util"
 	"github.com/andig/evcc/util/request"
+	"github.com/andig/evcc/vehicle/oidc"
 )
 
 const (
@@ -25,8 +24,7 @@ type Ford struct {
 	*embed
 	*request.Helper
 	user, password, vin string
-	token               string
-	tokenValid          time.Time
+	tokens              oidc.Tokens
 	chargeStateG        func() (float64, error)
 }
 
@@ -61,7 +59,8 @@ func NewFordFromConfig(other map[string]interface{}) (api.Vehicle, error) {
 
 	var err error
 	if cc.VIN == "" {
-		v.vin, err = findVehicle(v.vehicles())
+		// v.vin, err = findVehicle(v.vehicles())
+		v.vin, err = findVehicle(nil, nil)
 		if err == nil {
 			log.DEBUG.Printf("found vehicle: %v", v.vin)
 		}
@@ -84,35 +83,17 @@ func (v *Ford) login(user, password string) error {
 		return err
 	}
 
-	client := &http.Client{
-		Timeout:       v.Helper.Client.Timeout,
-		CheckRedirect: func(req *http.Request, via []*http.Request) error { return http.ErrUseLastResponse }, // don't follow redirects
+	var tokens oidc.Tokens
+	if err = v.DoJSON(req, &tokens); err == nil {
+		tokens.Valid = time.Now().Add(time.Duration(tokens.ExpiresIn) * time.Second)
+		v.tokens = tokens
 	}
 
-	resp, err := client.Do(req)
-	if err != nil {
-		return err
-	}
-
-	query, err := url.ParseQuery(resp.Header.Get("Location"))
-	if err != nil {
-		return err
-	}
-
-	token := query.Get("access_token")
-	expires, err := strconv.Atoi(query.Get("expires_in"))
-	if err != nil || token == "" || expires == 0 {
-		return errors.New("could not obtain token")
-	}
-
-	v.token = token
-	v.tokenValid = time.Now().Add(time.Duration(expires) * time.Second)
-
-	return nil
+	return err
 }
 
 func (v *Ford) request(uri string) (*http.Request, error) {
-	if v.token == "" || time.Since(v.tokenValid) > 0 {
+	if v.tokens.AccessToken == "" || time.Since(v.tokens.Valid) > 0 {
 		if err := v.login(v.user, v.password); err != nil {
 			return nil, err
 		}
@@ -121,48 +102,71 @@ func (v *Ford) request(uri string) (*http.Request, error) {
 	req, err := request.New(http.MethodGet, uri, nil, map[string]string{
 		"Content-type":   "application/json",
 		"Application-Id": "71A3AD0A-CF46-4CCF-B473-FC7FE5BC4592",
-		"Auth-Token":     v.token,
+		"Auth-Token":     v.tokens.AccessToken,
 	})
 
 	return req, err
 }
 
-// vehicles implements returns the list of user vehicles
-func (v *Ford) vehicles() ([]string, error) {
-	var br bmwVehiclesResponse
-	uri := fmt.Sprintf("%s/me/vehicles/v2/", bmwAPI)
+// // vehicles implements returns the list of user vehicles
+// func (v *Ford) vehicles() ([]string, error) {
+// 	var resp interface{}
+// 	uri := fmt.Sprintf("%s/api/vehicles/v4/status", fordAPI)
 
-	var vehicles []string
+// 	var vehicles []string
 
-	req, err := v.request(uri)
-	if err == nil {
-		err = v.DoJSON(req, &br)
-	}
+// 	req, err := v.request(uri)
+// 	if err == nil {
+// 		b, _ := httputil.DumpRequest(req, true)
+// 		fmt.Println(string(b))
+// 		err = v.DoJSON(req, &resp)
+// 	}
 
-	if err == nil {
-		for _, v := range br {
-			vehicles = append(vehicles, v.VIN)
-		}
-	}
+// 	if err == nil {
+// 		for _, v := range resp {
+// 			vehicles = append(vehicles, v.VIN)
+// 		}
+// 	}
 
-	return vehicles, err
-}
+// 	// return vehicles, err
+// 	return nil, err
+// }
 
 // chargeState implements the Vehicle.ChargeState interface
 func (v *Ford) chargeState() (float64, error) {
 	var resp struct {
-		Vehiclestatus interface{}
+		VehicleStatus struct {
+			Battery struct {
+				BatteryStatusActual struct {
+					Value int
+				}
+			}
+			BatteryFillLevel struct {
+				Value float64
+			}
+		}
+
+		// "chargingStatus": {
+		// 	"value": "EvseNotDetected",
+		// 	"status": "CURRENT",
+		// 	"timestamp": "10-27-2020 12:53:01"
+		// },
+		// "plugStatus": {
+		// 	"value": 1,
+		// 	"status": "CURRENT",
+		// 	"timestamp": "10-27-2020 12:53:01"
+		// },
+		// "batteryChargeStatus": null,
 	}
 
 	uri := fmt.Sprintf("%s/api/vehicles/v4/%s/status", fordAPI, v.vin)
 
 	req, err := v.request(uri)
-	if err != nil {
-		return 0, err
+	if err == nil {
+		err = v.DoJSON(req, &resp)
 	}
 
-	err = v.DoJSON(req, &resp)
-	return 0, err
+	return resp.VehicleStatus.BatteryFillLevel.Value, err
 }
 
 // ChargeState implements the Vehicle.ChargeState interface
