+++ vehicle/vw/identity.go
@@ -17,26 +17,6 @@ const (
 	OauthTokenURI = "https://mbboauth-1d.prd.ece.vwg-connect.com/mbbcoauth/mobile/oauth2/v1/token"
 )
 
-// Tokens is an OAuth tokens response
-type Tokens struct {
-	TokenType    string `json:"token_type"`
-	ExpiresIn    int    `json:"expires_in"`
-	IDToken      string `json:"id_token"`
-	AccessToken  string `json:"access_token"`
-	RefreshToken string `json:"refresh_token"`
-}
-
-// OIDCResponse is the well-known OIDC provider response
-// https://{oauth-provider-hostname}/.well-known/openid-configuration
-type OIDCResponse struct {
-	Issuer      string   `json:"issuer"`
-	AuthURL     string   `json:"authorization_endpoint"`
-	TokenURL    string   `json:"token_endpoint"`
-	JWKSURL     string   `json:"jwks_uri"`
-	UserInfoURL string   `json:"userinfo_endpoint"`
-	Algorithms  []string `json:"id_token_signing_alg_values_supported"`
-}
-
 // Identity provides the identity.vwgroup.io login
 type Identity struct {
 	*http.Client
